pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antspyt1w"
version = "0.9.3"
description = "T1w human neuroimage processing with antspyx"
readme = "README.md"
requires-python = ">=3.6"
license = { text = "Apache 2.0" }
authors = [
    { name = "Avants, Gosselin, Tustison", email = "stnava@gmail.com" }
]

dependencies = [
    "h5py>=3.10.0",
    "numpy>=1.19.4",
    "pandas>=1.0.1",
    "antspyx>=0.2.7",
    "antspynet>=0.2.5",
    "tensorflow-probability[tf]",
    "pathlib",
    "PyNomaly>=0.3.3",
    "pytest>=8.0.1",
    "scipy",
    "pillow",
    "matplotlib"
]

[project.urls]
homepage = "https://github.com/stnava/ANTsPyT1w"

[tool.setuptools.packages]
find = { include = ["antspyt1w"] }
